[chore] GameTimer
- 스크립트의 이름이 모호한 부분 수정
- 시간이 Update에서 이루어지지 않고, 코루틴으로 시간이 지나게 변경.
